Fix: Using JrSimpleTreeExtension to Provide Feature Context
