+++ jr-stree/src/main/java/com/fasterxml/jackson/jr/stree/JacksonJrsTreeCodec.java
@@ -18,7 +18,8 @@ public class JacksonJrsTreeCodec extends TreeCodec
 
     // @since 2.17
     protected boolean _failOnDuplicateKeys;
-    
+    protected boolean _useBigDecimalForDouble;
+
     public JacksonJrsTreeCodec() {
         this(null);
     }
@@ -31,7 +32,11 @@ public class JacksonJrsTreeCodec extends TreeCodec
     public void setFailOnDuplicateKeys(boolean state) {
         _failOnDuplicateKeys = state;
     }
-    
+
+    public void setUseBigDecimalForDouble(boolean state) {
+        _useBigDecimalForDouble = state;
+    }
+
     @SuppressWarnings("unchecked")
     @Override
     public <T extends TreeNode> T readTree(JsonParser p) throws IOException {
@@ -50,6 +55,9 @@ public class JacksonJrsTreeCodec extends TreeCodec
             return JrsBoolean.FALSE;
         case JsonTokenId.ID_NUMBER_INT:
         case JsonTokenId.ID_NUMBER_FLOAT:
+            if (_useBigDecimalForDouble) {
+                return new JrsNumber(p.getDecimalValue());
+            }
             return new JrsNumber(p.getNumberValue());
         case JsonTokenId.ID_STRING:
             return new JrsString(p.getText());
